Completed Part 3 of the Portfolio Project
